moving management functions into manage module
